Make linked data more case insensitive, move linked data opening hours parsing to LinkedDataParser
