+++ tests/test_ld_parser.py
@@ -2,6 +2,7 @@ import json
 
 from scrapy.http import HtmlResponse
 
+from locations.hours import OpeningHours
 from locations.linked_data_parser import LinkedDataParser
 
 
@@ -254,6 +255,322 @@ def test_ld_opening_hours_specification_as_list():
     )
 
     assert i["opening_hours"].as_opening_hours() == "Mo-Su 09:00-17:00"
+
+
+def test_ld_parse_opening_hours():
+    o = OpeningHours()
+    assert (
+        LinkedDataParser.parse_opening_hours(
+            json.loads(
+                """
+            {
+                "@context": "https://schema.org",
+                "@type": "Store",
+                "name": "Middle of Nowhere Foods",
+                "openingHoursSpecification":
+                [
+                    {
+                        "@type": "OpeningHoursSpecification",
+                        "dayOfWeek": [
+                            "http://schema.org/Monday",
+                            "https://schema.org/Tuesday",
+                            "Wednesday",
+                            "http://schema.org/Thursday",
+                            "http://schema.org/Friday"
+                        ],
+                        "opens": "09:00",
+                        "closes": "11:00"
+                    },
+                    {
+                        "@type": "OpeningHoursSpecification",
+                        "dayOfWeek": "http://schema.org/Saturday",
+                        "opens": "12:00",
+                        "closes": "14:00"
+                    }
+                ]
+            }
+            """
+            )
+        ).as_opening_hours()
+        == "Mo-Fr 09:00-11:00; Sa 12:00-14:00"
+    )
+
+
+def test_ld_parse_opening_hours_string():
+    o = OpeningHours()
+    assert (
+        LinkedDataParser.parse_opening_hours(
+            json.loads(
+                """
+            {
+                "@context": "https://schema.org",
+                "@type": "Pharmacy",
+                "name": "Philippa's Pharmacy",
+                "description": "A superb collection of fine pharmaceuticals for your beauty and healthcare convenience, a department of Delia's Drugstore.",
+                "openingHours": "Mo,Tu,We,Th 09:00-12:00",
+                "telephone": "+18005551234"
+            }
+            """
+            )
+        ).as_opening_hours()
+        == "Mo-Th 09:00-12:00"
+    )
+
+    o = OpeningHours()
+    assert (
+        LinkedDataParser.parse_opening_hours(
+            json.loads(
+                """
+            {
+                "@context": "https://schema.org",
+                "@type": "Pharmacy",
+                "name": "Philippa's Pharmacy",
+                "description": "A superb collection of fine pharmaceuticals for your beauty and healthcare convenience, a department of Delia's Drugstore.",
+                "openingHours": "Mo-Th 09:00-12:00",
+                "telephone": "+18005551234"
+            }
+            """
+            )
+        ).as_opening_hours()
+        == "Mo-Th 09:00-12:00"
+    )
+
+    o = OpeningHours()
+    assert (
+        LinkedDataParser.parse_opening_hours(
+            json.loads(
+                """
+            {
+                "@context": "https://schema.org",
+                "@type": "Pharmacy",
+                "name": "Philippa's Pharmacy",
+                "description": "A superb collection of fine pharmaceuticals for your beauty and healthcare convenience, a department of Delia's Drugstore.",
+                "openingHours": "Mo-Tu 09:00-12:00 We,Th 09:00-12:00",
+                "telephone": "+18005551234"
+            }
+            """
+            )
+        ).as_opening_hours()
+        == "Mo-Th 09:00-12:00"
+    )
+
+
+def test_ld_parse_opening_hours_days_3_chars():
+    o = OpeningHours()
+    assert (
+        LinkedDataParser.parse_opening_hours(
+            json.loads(
+                """
+            {
+                "@context": "https://schema.org",
+                "@type": "Pharmacy",
+                "name": "Philippa's Pharmacy",
+                "description": "A superb collection of fine pharmaceuticals for your beauty and healthcare convenience, a department of Delia's Drugstore.",
+                "openingHours": "Mon-Thu 09:00-12:00",
+                "telephone": "+18005551234"
+            }
+            """
+            )
+        ).as_opening_hours()
+        == "Mo-Th 09:00-12:00"
+    )
+
+    o = OpeningHours()
+    assert (
+        LinkedDataParser.parse_opening_hours(
+            json.loads(
+                """
+            {
+                "@context": "https://schema.org",
+                "@type": "Pharmacy",
+                "name": "Philippa's Pharmacy",
+                "description": "A superb collection of fine pharmaceuticals for your beauty and healthcare convenience, a department of Delia's Drugstore.",
+                "openingHours": "Mon-Tue 09:00-12:00 Wed,Thu 09:00-12:00",
+                "telephone": "+18005551234"
+            }
+            """
+            )
+        ).as_opening_hours()
+        == "Mo-Th 09:00-12:00"
+    )
+
+    o = OpeningHours()
+    assert (
+        LinkedDataParser.parse_opening_hours(
+            json.loads(
+                """
+            {
+                "@context": "https://schema.org",
+                "@type": "Pharmacy",
+                "openingHours": "Mon-Sat 10:00 - 19:00 Sun 12:00-17:00"
+            }
+            """
+            )
+        ).as_opening_hours()
+        == "Mo-Sa 10:00-19:00; Su 12:00-17:00"
+    )
+
+
+def test_ld_parse_opening_hours_array():
+    o = OpeningHours()
+    assert (
+        LinkedDataParser.parse_opening_hours(
+            json.loads(
+                """
+            {
+                "@context": "https://schema.org",
+                "@type": ["TouristAttraction", "AmusementPark"],
+                "name": "Disneyland Paris",
+                "description": "It's an amusement park in Marne-la-Vallée, near Paris, in France and is the most visited theme park in all of France and Europe.",
+                "openingHours":["Mo-Fr 10:00-19:00", "Sa 10:00-22:00", "Su 10:00-21:00"],
+                "isAccessibleForFree": false,
+                "currenciesAccepted": "EUR",
+                "paymentAccepted":"Cash, Credit Card",
+                "url":"http://www.disneylandparis.it/"
+            }
+            """
+            )
+        ).as_opening_hours()
+        == "Mo-Fr 10:00-19:00; Sa 10:00-22:00; Su 10:00-21:00"
+    )
+
+
+def test_ld_parse_opening_hours_day_range():
+    o = OpeningHours()
+    assert (
+        LinkedDataParser.parse_opening_hours(
+            json.loads(
+                """
+            {
+                "@context": "https://schema.org",
+                "openingHours": ["Th-Tu 09:00-17:00"]
+            }
+            """
+            )
+        ).as_opening_hours()
+        == "Mo-Tu 09:00-17:00; Th-Su 09:00-17:00"
+    )
+
+
+def test_ld_parse_opening_hours_array_with_commas():
+    o = OpeningHours()
+    assert (
+        LinkedDataParser.parse_opening_hours(
+            json.loads(
+                """
+            {
+                "@context": "https://schema.org",
+                "openingHours": ["Mo-Su 00:00-01:00, 04:00-00:00"]
+            }
+            """
+            )
+        ).as_opening_hours()
+        == "Mo-Su 00:00-01:00,04:00-24:00"
+    )
+
+
+def test_ld_parse_opening_hours_closed():
+    o = OpeningHours()
+    assert (
+        LinkedDataParser.parse_opening_hours(
+            json.loads(
+                """
+            {
+                "@context": "https://schema.org",
+                "openingHours": [
+                    "Mo Closed",
+                    "Tu Closed",
+                    "We Closed",
+                    "Th Closed",
+                    "Fr Closed",
+                    "Sa Closed",
+                    "Su Closed"
+                ]
+            }
+            """
+            )
+        ).as_opening_hours()
+        == "Mo-Su closed"
+    )
+
+
+def test_ld_parse_opening_hours_closed_range():
+    o = OpeningHours()
+    assert (
+        LinkedDataParser.parse_opening_hours(
+            json.loads(
+                """
+            {
+                "@context": "https://schema.org",
+                "openingHours": ["Mo-Su Closed"]
+            }
+            """
+            )
+        ).as_opening_hours()
+        == "Mo-Su closed"
+    )
+
+
+def test_ld_parse_opening_hours_no_commas():
+    o = OpeningHours()
+    assert (
+        LinkedDataParser.parse_opening_hours(
+            json.loads(
+                """
+            {
+                "@context": "https://schema.org",
+                "openingHours": "Su 07:00 - 23:00 Mo 07:00 - 23:00 Tu 07:00 - 23:00 We 07:00 - 23:00 Th 07:00 - 23:00 Fr 07:00 - 23:00 Sa 07:00 - 23:00 "
+            }
+            """
+            )
+        ).as_opening_hours()
+        == "Mo-Su 07:00-23:00"
+    )
+
+
+def test_ld_parse_opening_hours_no_commas_closed():
+    o = OpeningHours()
+    assert (
+        LinkedDataParser.parse_opening_hours(
+            json.loads(
+                """
+            {
+                "@context": "https://schema.org",
+                "openingHours": "Su closed Mo closed Tu closed We closed Th closed Fr closed Sa closed "
+            }
+            """
+            )
+        ).as_opening_hours()
+        == "Mo-Su closed"
+    )
+
+
+def test_ld_parse_time_format():
+    o = OpeningHours()
+    assert (
+        LinkedDataParser.parse_opening_hours(
+            json.loads(
+                """
+            {
+                "@context": "https://schema.org",
+                "@type": "Store",
+                "name": "Middle of Nowhere Foods",
+                "openingHoursSpecification":
+                [
+                    {
+                        "@type": "OpeningHoursSpecification",
+                        "dayOfWeek": "http://schema.org/Saturday",
+                        "opens": "12:00:00",
+                        "closes": "14:00:00"
+                    }
+                ]
+            }
+            """
+            ),
+            "%H:%M:%S",
+        ).as_opening_hours()
+        == "Sa 12:00-14:00"
+    )
 
 
 def test_ld_lat_lon():
